requirements.txt
--index-url https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple

openai
transformers
torch
aiohttp
websockets
python-dotenv
fastapi
uvicorn
sentence-transformers
chromadb
python-multipart
pyqt6
mss

# Gemini
google.generativeai

#memory-server
mem0ai >= 0.1.108
langchain >= 0.3.25
langchain-community >= 0.3.25
langchain-core >= 0.3.65
langchain-neo4j >= 0.4.0
rank-bm25 >= 0.2.2
mcp >= 0.9.0
fastmcp >= 0.2.0